M2a: Snapshot wire format — proto3 schema, codec, content digests (#6)
* M2a: snapshot wire format — proto3 schema, codec, content digests

The snapshot is a contract between a Python control plane that compiles it and
a Go data plane that serves from it, so it needs one definition generated for
both rather than two hand-maintained ones that drift.

  - proto/gateway/v1/snapshot.proto is the schema. Layers are transferable
    individually, which is the point of the layering: a budget edit ships one
    small tenant layer rather than the whole catalog.
  - internal/wire converts between the schema and the domain types. Conversion
    goes through core's Spec types, never into or out of a built layer, so the
    direction of travel is explicit: bytes -> spec -> validated layer.
  - Content digests. A layer is hashed with its own digest field cleared, over
    deterministically marshalled bytes, so a worker can skip a re-fetch it
    already holds and two workers claiming a version can be proven to agree.
    This is a corruption check, not an authenticity one — anyone who can
    rewrite the layer can rewrite the digest. Authenticity is the transport's.

Validation stays in core. The wire layer shapes data and core decides whether
it is coherent, so there is one definition of a valid snapshot rather than a
second, weaker one here.

Enums convert through explicit tables rather than numeric casts. The numbering
agrees today; a cast would make that coincidence load-bearing, so reordering a
domain constant would silently change what a stored snapshot means. An unknown
enum value fails the layer instead of defaulting to the most permissive one.

Two bugs the round-trip tests caught, fixed at the source:

  - core.Wrap returned *Error, so `return core.Wrap(...)` into an `error`
    result produced a non-nil interface holding a nil pointer: an error that is
    not nil and has no message. It now returns `error`, removing the trap for
    every caller rather than patching the one site that hit it.
  - Deployment.Weight was uint8 and HeadroomBasisPoints uint16, both narrower
    than the wire. An out-of-range value truncated into a plausible one — 300
    became 44 — instead of failing core's range check. Both widened to match.

Also: proto3 rather than editions, since protoc 25.x treats editions as
experimental. Generated code is committed so the build needs no toolchain, and
CI regenerates it to prove it still matches the schema.


* ci: make the generated-code drift check version-independent

The check failed on the protoc version banner rather than on real drift: the
generated file records which protoc built it, and CI's protoc is not the
developer's. Comparing that line means the check reports a false positive
whenever the two differ, which is most of the time.

  - make proto strips the version banner, so the committed artifact depends on
    the schema rather than on whose machine ran the generator.
  - CI pins protoc to 25.3 and installs protoc-gen-go at the protobuf version
    the module already depends on, rather than @latest, so the generator itself
    cannot drift underneath the check.

The check now fails only when the .proto and the committed Go disagree, which
is the thing it was for.


* ci: quote the command substitution in the protoc-gen-go install

shellcheck SC2046: an unquoted $(...) is word-split, so a version string with
a space would silently become two arguments. Caught by actionlint via reviewdog
on the previous commit — which is the automated reviewer doing its job.


---------

+++ dataplane/internal/core/budget.go
@@ -50,8 +50,9 @@ type BudgetState struct {
 
 	// HeadroomBasisPoints holds back a slice of a hard limit so that requests
 	// already streaming when the budget tips can finish without overshooting.
-	// 500 = 5%, which is the documented default.
-	HeadroomBasisPoints uint16
+	// 500 = 5%, which is the documented default. uint32 to match the wire
+	// format, for the same reason Deployment.Weight is.
+	HeadroomBasisPoints uint32
 }
 
 // Available is the spend remaining before the headroom band begins.
